# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HR Analytics Summit Conference Registration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings/{bookingId}/attendees:
    post:
      operationId: createAttendee
      tags: [attendees]
      summary: Add an attendee to a booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendeeBody"
      responses:
        "201":
          description: Attendee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{bookingId}/attendees/{attendeeId}:
    patch:
      operationId: updateAttendee
      tags: [attendees]
      summary: Update an attendee's details
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
        - name: attendeeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttendeeBody"
      responses:
        "200":
          description: Attendee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== PRICING ====================
...
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: x-admin-token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Attendee:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        isLead:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
        company:
          type: string
        workEmail:
          type: string
        phone:
          type: ["string", "null"]
        isTbc:
          type: boolean
        gdprConsent:
          type: boolean
        gdprConsentAt:
          type: ["string", "null"]
          format: date-time
        dietaryAccessibility:
          type: ["string", "null"]
        seatIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - bookingId
        - isLead
        - isTbc
        - firstName
        - lastName
        - jobTitle
        - company
        - workEmail
        - gdprConsent
        - createdAt
        - updatedAt
    CreateAttendeeBody:
      type: object
      properties:
        isLead:
          type: boolean
        isTbc:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
        company:
          type: string
        workEmail:
          type: string
        phone:
          type: ["string", "null"]
        dietaryAccessibility:
          type: ["string", "null"]
        gdprConsent:
          type: boolean
        seatIndex:
          type: integer
      required:
        - gdprConsent
    UpdateAttendeeBody:
      type: object
      properties:
        isTbc:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
        company:
          type: string
        workEmail:
          type: string
        phone:
          type: ["string", "null"]
        dietaryAccessibility:
          type: ["string", "null"]
        gdprConsent:
          type: boolean

    # ==================== PRICING SCHEMAS ====================
...